Break kinship-chart.ts into multiple files
